requirements.txt
matplotlib==3.8.2
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0
snowflake_connector_python==3.5.0
streamlit==1.28.2
